+++ src/marshmallow/fields.py
@@ -525,8 +525,17 @@ class Nested(Field):
                 nested = self.nested
 
             if isinstance(nested, SchemaABC):
-                self._schema = nested
+                self._schema = copy.deepcopy(nested)
                 self._schema.context.update(context)
+                # Respect only and exclude passed from parent and re-initialize fields
+                set_class = self._schema.set_class
+                if self.only is not None:
+                    original = self._schema.only
+                    self._schema.only = set_class(self.only).intersection(original)
+                if self.exclude:
+                    original = self._schema.exclude
+                    self._schema.exclude = set_class(self.exclude).union(original)
+                self._schema._init_fields()
             else:
                 if isinstance(nested, type) and issubclass(nested, SchemaABC):
                     schema_class = nested
@@ -570,17 +579,19 @@ class Nested(Field):
         schema = self.schema
         if nested_obj is None:
             return None
+        many = schema.many or self.many or many
         return schema.dump(nested_obj, many=self.many or many)
 
     def _test_collection(self, value, many=False):
-        many = self.many or many
+        many = self.schema.many or self.many or many
         if many and not utils.is_collection(value):
             raise self.make_error("type", input=value, type=value.__class__.__name__)
 
     def _load(self, value, data, partial=None, many=False):
+        many = self.schema.many or self.many or many
         try:
             valid_data = self.schema.load(
-                value, unknown=self.unknown, partial=partial, many=self.many or many
+                value, unknown=self.unknown, partial=partial, many=many
             )
         except ValidationError as error:
             raise ValidationError(
@@ -862,7 +873,7 @@ class UUID(String):
 class Number(Field):
     """Base class for number fields.
 
-    :param bool as_string: If True, format the serialized value as a string.
+    :param bool as_string: If `True`, format the serialized value as a string.
     :param kwargs: The same keyword arguments that :class:`Field` receives.
     """
 
@@ -911,6 +922,8 @@ class Number(Field):
 class Integer(Number):
     """An integer field.
 
+    :param bool strict: If `True`, only integer types are valid.
+        Otherwise, any value castable to `int` is valid.
     :param kwargs: The same keyword arguments that :class:`Number` receives.
     """
 
@@ -937,7 +950,7 @@ class Float(Number):
 
     :param bool allow_nan: If `True`, `NaN`, `Infinity` and `-Infinity` are allowed,
         even though they are illegal according to the JSON specification.
-    :param bool as_string: If True, format the value as a string.
+    :param bool as_string: If `True`, format the value as a string.
     :param kwargs: The same keyword arguments that :class:`Number` receives.
     """
 
@@ -984,11 +997,11 @@ class Decimal(Number):
     :param int places: How many decimal places to quantize the value. If `None`, does
         not quantize the value.
     :param rounding: How to round the value during quantize, for example
-        `decimal.ROUND_UP`. If None, uses the rounding value from
+        `decimal.ROUND_UP`. If `None`, uses the rounding value from
         the current thread's context.
     :param bool allow_nan: If `True`, `NaN`, `Infinity` and `-Infinity` are allowed,
         even though they are illegal according to the JSON specification.
-    :param bool as_string: If True, serialize to a string instead of a Python
+    :param bool as_string: If `True`, serialize to a string instead of a Python
         `decimal.Decimal` type.
     :param kwargs: The same keyword arguments that :class:`Number` receives.